pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cae-mesh-automation"
version = "0.1.0"
description = "Deterministic AI-assisted CAE mesh automation with CDF and AMG workflows."
requires-python = ">=3.11"
dependencies = [
  "jsonschema>=4",
  "numpy>=1.24",
  "pandas>=2",
  "pyarrow>=12",
  "PyYAML>=6",
  "torch>=2",
]

[project.scripts]
cdf = "cae_dataset_factory.cli:main"
amg = "ai_mesh_generator.cli:main"
train-brep-assembly-net = "training_pipeline.train:main"
evaluate-brep-assembly-net = "training_pipeline.evaluate:main"
export-amg-model = "training_pipeline.export_model:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider --basetemp=./pytest_tmp2"
